Refactoring all the tests
