--- generic/piranha-tree-sitter/test-resources/java/feature_flag_system_1/treated/configurations/edges.toml
@@ -1,27 +1,0 @@
-# Copyright (c) 2022 Uber Technologies, Inc.
-# 
-# <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
-# except in compliance with the License. You may obtain a copy of the License at
-# <p>http://www.apache.org/licenses/LICENSE-2.0
-# 
-# <p>Unless required by applicable law or agreed to in writing, software distributed under the
-# License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
-# express or implied. See the License for the specific language governing permissions and
-# limitations under the License.
-
-# The edges in this file specify the the flow between the rules
-# [[edges]]
-# scope = "Global"
-# from = "delete_flag_method_declaration_fully_annotated"
-# to = [
-#   "replace_flag_method_with_boolean_literal",
-#   "delete_put",
-# ]
-
-# [[edges]]
-# scope = "Global"
-# from = "delete_flag_method_declaration_inside_annotated_interface"
-# to = [
-#   "replace_flag_method_with_boolean_literal",
-#   "delete_put",
-# ]
